requirements.txt
beautifulsoup4==4.9.0
certifi==2020.4.5.1
chardet==3.0.4
Django==2.2.12
django-appconf==1.0.4
django-autoslug==1.9.7
django-bootstrap4==1.1.1
django-mathfilters==1.0.0
gunicorn==20.0.4
idna==2.9
psycopg2==2.8.5
pytz==2020.1
requests==2.23.0
soupsieve==2.0
sqlparse==0.3.1
urllib3==1.25.9
